bcm2835_rng_rust: add support for `DrvData`
To demonstrate device driver data in a non-trivial way, create
a `miscdev` for each discovered device. This `miscdev` can only be
opened and read. When read from userspace, it returns four zero bytes
at a time.

Note that `DrvData` consists of a `Pin<Box<miscdev::Registration>>`,
which is a pinned structure. This demonstrates that pinned or self-
referential structures may be used in `DrvData`.

Signed-off-by: Sven Van Asbroeck <thesven73@gmail.com>