rust/kernel: use zero-cost method to construct certain `Error`s
In a previous PR, the type invariant for `Error` is enforced using
a runtime check. This is non-zero cost.

However we may decide to trust the return value of certain kernel C
functions. In such cases, no runtime check is required to enforce the
type invariant. So we can return to zero-cost.

This patch removes invariant checks from kernel C functions that
return 0 on success, or a non-zero errno on failure.

Signed-off-by: Sven Van Asbroeck <thesven73@gmail.com>