rust/kernel: remove redundant `Error::from_kernel_errno()`
It turns out that we don't need an `Error` constructor that
checks the `errno` invariant at runtime. This is because we
trust return values originating from kernel C. And any return
values originating from Rust code can be constucted using
`Error::` constants, which also do not need a runtime check:

```rust
    return Err(Error::EBUSY);
```

If we find we do require this function in the future, we may
restore it simply by reverting this commit.

Signed-off-by: Sven Van Asbroeck <thesven73@gmail.com>