platdev: introduce `shutdown` callback
Before proceeding with the trickier aspects of platdev (and the
driver model in general), add a `PlatformDriver` callback which
accesses its `PlatData` as a `Borrowed`.

The majority of `bindings::platform_driver` callbacks access their
platform data by reference. So it would be prudent to have at least
one callback that demonstrates this behaviour.

We pick `shutdown` because it is trivial to implement, and easy to
trigger by executing `shutdown -P now` in the shell.

Signed-off-by: Sven Van Asbroeck <thesven73@gmail.com>